Entities and migrations moved to DAL project
