+++ WebAPI/DAL/Entities/Following.cs
@@ -5,16 +5,19 @@ using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebAPI.Models
+namespace DAL.Entities
 {
     public class Following
     {
         public int Id { get; set; }
 
         public int? FollowerProfileId { get; set; }
+
         [ForeignKey("FollowerProfileId")]
         public Profile FollowerProfile { get; set; }
+
         public int? FollowingProfileId { get; set; }
+
         [ForeignKey("FollowingProfileId")]
         public Profile FollowingProfile { get; set; }
