+++ WebAPI/DAL/Entities/Like.cs
@@ -5,16 +5,20 @@ using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebAPI.Models
+namespace DAL.Entities
 {
     public class Like
     {
         public int Id { get; set; }
+
         public int? ProfileId { get; set; }
+
         [ForeignKey("ProfileId")]
         public Profile Profile { get; set; }
+
         [Required]
         public int PostId { get; set; }
+
         [ForeignKey("PostId")]
         public Post Post { get; set; }
     }
