+++ WebAPI/DAL/Entities/Post.cs
@@ -5,24 +5,27 @@ using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebAPI.Models
+namespace DAL.Entities
 {
     public class Post
     {
         public int Id { get; set; }
+
         [Required]
         [MaxLength(140)]
         public string PostText { get; set; }
+
         [Required]
         public DateTime PostingDate { get; set; }
 
         [Required]
         public int ProfileId { get; set; }
+
         [ForeignKey("ProfileId")]
         public Profile Profile { get; set; }
 
         public List<Reply> Replies { get; set; }
-        public List<Like> Likes { get; set; }
 
+        public List<Like> Likes { get; set; }
     }
 }
