+++ WebAPI/DAL/Entities/Reply.cs
@@ -5,22 +5,27 @@ using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebAPI.Models
+namespace DAL.Entities
 {
     public class Reply
     {
         public int Id { get; set; }
+
         [Required]
         [MaxLength(140)]
         public string ReplyText { get; set; }
+
         [Required]
         public DateTime ReplyingDate { get; set; }
 
         public int? ProfileId { get; set; }
+
         [ForeignKey("ProfileId")]
         public Profile Profile { get; set; }
+
         [Required]
         public int PostId { get; set; }
+
         [ForeignKey("PostId")]
         public Post Post { get; set; }
     }
