+++ WebAPI/DAL/Entities/User.cs
@@ -4,17 +4,20 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace WebAPI.Models
+namespace DAL.Entities
 {
     public class User
     {
         public int Id { get; set; }
+
         [Required]
         [MaxLength(256)]
         public string UserName { get; set; }
+
         [Required]
         [MaxLength(256)]
         public string Email { get; set; }
+
         [Required]
         [MaxLength(256)]
         public string Password { get; set; }
