+++ WebAPI/WebAPI.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31424.327
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebAPI", "WebAPI\WebAPI.csproj", "{4D9B8855-94E3-41D2-BF45-54B763EEA4F1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebAPI", "WebAPI\WebAPI.csproj", "{4D9B8855-94E3-41D2-BF45-54B763EEA4F1}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DAL", "DAL\DAL.csproj", "{6AADFB81-DC27-44A3-8AF6-D4388581CB20}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{4D9B8855-94E3-41D2-BF45-54B763EEA4F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4D9B8855-94E3-41D2-BF45-54B763EEA4F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4D9B8855-94E3-41D2-BF45-54B763EEA4F1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6AADFB81-DC27-44A3-8AF6-D4388581CB20}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6AADFB81-DC27-44A3-8AF6-D4388581CB20}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6AADFB81-DC27-44A3-8AF6-D4388581CB20}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6AADFB81-DC27-44A3-8AF6-D4388581CB20}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
